requirements.txt
allure-pytest==2.8.13
allure-python-commons==2.8.13
Pillow==7.1.2
pytest==5.4.2
requests==2.23.0
requests-oauthlib==1.3.0
selenium==4.0.0a5
unittest2==1.1.0
urllib3==1.25.9
webm==0.12.1
opencv-python==4.2.0.34
pytest-xdist==1.32.0